Scale estimation fix (#1)
* Merging changes from Horizon0156:master

* Start doc with install instructions using catkin

* Doc of drone_stateestimation + quick start

* some fixes in doc

* some fixes in doc

* updated thirdparty to latest libcvd

* Doc for drone_autopilot

* Doc for drone_gui

* Doc: add tricks

The doc is more or less done.
However, a lot of things have been copied from .txt file or ROS wiki.
Should be checked

* Delete readme_gui.txt

* Delete readme_autopilot.txt

* Delete readme_tricks.txt

* fix some compile warnings

* prepare services by creating functions

* specify services

* implement services

* ignoring temp files

* services: rename Dist into Distance

* fix memory error

* compiled TooN with gcc 4.8.2

* Fix scale estimation calculation

It incorrectly used the kalman filter to retrieve the previous
visual field position.  This set up a circular dependency between
the scale parameter and the filter.

The fix is to use the previous visual field position as specified
in the associated papers.